Refine graph/ header file (#4459)
