--- oneflow/core/graph/foreign_input_compute_task_node.h
@@ -1,34 +1,0 @@
-/*
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-#ifndef ONEFLOW_CORE_GRAPH_FOREIGN_INPUT_COMPUTE_TASK_NODE_H_
-#define ONEFLOW_CORE_GRAPH_FOREIGN_INPUT_COMPUTE_TASK_NODE_H_
-
-#include "oneflow/core/graph/foreign_io_compute_task_node.h"
-
-namespace oneflow {
-
-class ForeignInputCompTaskNode final : public ForeignIOCompTaskNode {
- public:
-  OF_DISALLOW_COPY_AND_MOVE(ForeignInputCompTaskNode);
-  ForeignInputCompTaskNode() = default;
-  ~ForeignInputCompTaskNode() override = default;
-
-  TaskType GetTaskType() const override { return TaskType::kForeignInput; }
-};
-
-}  // namespace oneflow
-
-#endif  // ONEFLOW_CORE_GRAPH_FOREIGN_INPUT_COMPUTE_TASK_NODE_H_
