--- oneflow/core/graph/tick_compute_task_node.h
@@ -1,41 +1,0 @@
-/*
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-#ifndef ONEFLOW_CORE_GRAPH_TICK_COMPUTE_TASK_NODE_H_
-#define ONEFLOW_CORE_GRAPH_TICK_COMPUTE_TASK_NODE_H_
-
-#include "oneflow/core/graph/compute_task_node.h"
-
-namespace oneflow {
-
-class TickCompTaskNode final : public CompTaskNode {
- public:
-  OF_DISALLOW_COPY_AND_MOVE(TickCompTaskNode);
-  TickCompTaskNode() = default;
-  ~TickCompTaskNode() = default;
-
-  bool IsMeaningLess() override { return false; }
-  TaskType GetTaskType() const override { return TaskType::kTick; }
-
- private:
-  void ProduceAllRegstsAndBindEdges() override;
-  void ConsumeAllRegsts() override;
-  void BuildExecGphAndRegst() override;
-  bool IsIndependent() const override { return true; }
-};
-
-}  // namespace oneflow
-
-#endif  // ONEFLOW_CORE_GRAPH_TICK_COMPUTE_TASK_NODE_H_
