+++ oneflow/core/graph_impl/callback_notify_compute_task_node.cpp
@@ -13,9 +13,40 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/graph/callback_notify_compute_task_node.h"
+#include "oneflow/core/graph/compute_task_node.h"
 
 namespace oneflow {
+
+class CallbackNotifyCompTaskNode final : public CompTaskNode {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(CallbackNotifyCompTaskNode);
+  CallbackNotifyCompTaskNode() = default;
+  ~CallbackNotifyCompTaskNode() = default;
+
+  TaskType GetTaskType() const override { return TaskType::kCallbackNotify; }
+  bool IsIndependent() const override { return true; }
+
+ private:
+  void ProduceAllRegstsAndBindEdges() override;
+  void ConsumeAllRegsts() override;
+  void BuildExecGphAndRegst() override;
+};
+
+void CallbackNotifyCompTaskNode::ProduceAllRegstsAndBindEdges() {}
+
+void CallbackNotifyCompTaskNode::ConsumeAllRegsts() {
+  ForEachInDataEdge([&](TaskEdge* edge) { ConsumeRegst("in", edge->GetSoleRegst()); });
+}
+
+void CallbackNotifyCompTaskNode::BuildExecGphAndRegst() {
+  ExecNode* node = mut_exec_gph().NewNode();
+  node->mut_op() = this->op();
+  for (const std::string& ibn : node->op()->input_bns()) {
+    node->BindBnWithOneOfTheRegsts(ibn, GetConsumedRegst("in"));
+  }
+  CHECK(node->op()->tmp_bns().empty());
+  CHECK(node->op()->output_bns().empty());
+}
 
 REGISTER_INDEPENDENT_THREAD_NUM(TaskType::kCallbackNotify, 1);
