+++ oneflow/core/graph_impl/wait_and_send_ids_compute_task_node.cpp
@@ -13,9 +13,27 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/graph/wait_and_send_ids_compute_task_node.h"
+#include "oneflow/core/graph/compute_task_node.h"
 
 namespace oneflow {
+
+class WaitAndSendIdsCompTaskNode final : public CompTaskNode {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(WaitAndSendIdsCompTaskNode);
+  WaitAndSendIdsCompTaskNode() = default;
+  ~WaitAndSendIdsCompTaskNode() override = default;
+
+  void ProduceAllRegstsAndBindEdges() override;
+  void ConsumeAllRegsts() override {}
+  void BuildExecGphAndRegst() override;
+  bool IsMeaningLess() override { return false; }
+
+  TaskType GetTaskType() const override { return TaskType::kWaitAndSendIds; }
+  bool IsIndependent() const override { return true; }
+
+ private:
+  void InferProducedDataRegstTimeShape() override;
+};
 
 void WaitAndSendIdsCompTaskNode::ProduceAllRegstsAndBindEdges() {
   std::shared_ptr<RegstDesc> out_regst = ProduceRegst("out", false, 100, 100);
